requirements.txt
git+https://github.com/vahee/agentopy@main
openai
requests
numpy
selenium
duckduckgo_search
beautifulsoup4
pyyaml
pycodestyle
dacite
imap_tools
pytest
lxml
tiktoken
orjson==3.8.10
torch
pytest-asyncio
